backend/requirements.txt
fastapi==0.124.4
uvicorn==0.33.0
python-dotenv==1.0.1
supabase==2.6.0
httpx==0.27.2
requests==2.32.4
